Stop erroring when there's no new news
The news scraper fails when there's no new news from the scraper. Since we expect that to be common, these failures can be confusing and, even worse, can mask actual, meaningful failures. Instead, check for whether there was new news and only try to commit if so -- that way we don't cause errors if there's nothing new.

This also breaks up the workflow into more granular steps, so it's easier to see where exactly something is going wrong.

Finally, it changes the commit message to include information about what commit of sfbrigade/data-covid19-sfbayarea was used to produce the data.