Proposed FAQ Content to JSON automator
First go at a proposed faq-json automator. Have included plenty of comments so team members who may not be native to Python can see what's going on should they need to. The "faq.json" output file points to the file in the repo, so be careful when testing! You can always change the name of the output file in the second to last line while assessing ;-)