Create PR on partial news scraper failures
It’s possible for the news scraper, like the data scraper, to succeed on some counties while failing on others. In this case, it exits with an error, but still did useful work.

Instead of stopping the workflow in its tracks, this change allows the workflow to continue so we still get a PR with updates from the successful counties.

Like the data scraper, it also sends error logs to Slack. Unlike the data scraper, it also fails the workflow show it shows up as an error in the actions panel on GitHub. If this works well (it's the first time I've used the conditional features in actions), I'll port this back over to the data scraper, too.