Fix old and new data not getting merged safely
In opening the data file for both reading and writing at the same time, we would fail to properly write new output. This resulted in issues like the San Mateo data disappearing entirely if the San Mateo scraper failed instead of just having the San Mateo data not get updated.