Fail the scraper workflow if the scraper errored
We handle errors while scraping so that, if only some counties fail, we can continue on to create a PR for the successfully scraped counties. However, we should still signal an overall failure afterwards. This change now does so, similar to how we handle the same situation in the news workflow.